Modificat el tipus de la variable diners de Float a BigDecimal, amb dos decimals,

+++ src/main/java/com/dualion/controldiners/web/rest/vm/ExtreureVM.java
@@ -1,6 +1,7 @@
 package com.dualion.controldiners.web.rest.vm;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 import javax.validation.constraints.NotNull;
 
@@ -11,17 +12,17 @@ public class ExtreureVM implements Serializable {
 	private static final long serialVersionUID = 1L;
 	
 	@NotNull
-	Float diners;
+	BigDecimal diners;
 
 	@JsonCreator
 	public ExtreureVM() {
 	}
 
-	public Float getDiners() {
+	public BigDecimal getDiners() {
 		return diners;
 	}
 
-	public void setDiners(Float diners) {
+	public void setDiners(BigDecimal diners) {
 		this.diners = diners;
 	}
